+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Query/Pipeline/FullPipelineTests.cs
@@ -367,9 +367,8 @@ namespace Microsoft.Azure.Cosmos.Tests.Query.Pipeline
                 "HelloWorld",
                 It.IsAny<IReadOnlyList<Documents.Routing.Range<string>>>(),
                 It.IsAny<bool>(),
-                It.IsAny<ITrace>(),
-                It.IsAny<CancellationToken>()))
-                .Returns((string resourceLink, string collectionResourceId, IReadOnlyList<Documents.Routing.Range<string>> providedRanges, bool forceRefresh, ITrace trace, CancellationToken ct) => Task.FromResult(partitionKeyRanges));
+                It.IsAny<ITrace>()))
+                .Returns((string resourceLink, string collectionResourceId, IReadOnlyList<Documents.Routing.Range<string>> providedRanges, bool forceRefresh, ITrace trace) => Task.FromResult(partitionKeyRanges));
 
             mockClient.Setup(x => x.TryGetPartitionedQueryExecutionInfoAsync(
                 It.IsAny<SqlQuerySpec>(),